# Objective-C printouts and loops performance