I have:

* [X] Read the project [README](../README.md), including the [benchmark descriptions](../README.md#available-benchmarks)

## Description of changes

I set out to fix the prints from fibonacci and loops, which failed `check.sh`. So that's done, but then the loops performance was really awful, which didn't match my perception about Objective-C at all. Using a C array fixed it. I guess in most ObjC programs `NSMutableArray` is what you go for, unless you are really going to crunch through the array, then you probably pick something more suiting. 😄 


### Loops performance

**Before**

```
Checking Objective-C
Check passed
Benchmarking Objective-C
Benchmark 1: ./objc/code 40
  Time (mean ± σ):     13.329 s ±  0.279 s    [User: 12.921 s, System: 0.088 s]
  Range (min … max):   13.043 s … 13.800 s    7 runs
```

**After**

```
Checking Objective-C
Check passed
Benchmarking Objective-C
Benchmark 1: ./objc/code 40
  Time (mean ± σ):     518.4 ms ±   3.6 ms    [User: 508.3 ms, System: 4.7 ms]
  Range (min … max):   513.1 ms … 522.7 ms    7 runs
```